requirements.txt
autopep8
Django
django-import-export
mysqlclient
django-admin-interface
django-admin-rangefilter